refactor: CrewAttendanceStatus의 출석 상태 관리 책임 CrewAttendance에게 이전
