+++ src/main/java/domain/CrewAttendance.java
@@ -4,17 +4,30 @@ import java.time.LocalTime;
 import java.util.Objects;
 
 public class CrewAttendance {
-    
+
     private final AttendanceTime attendanceTime;
-    private final CrewAttendanceStatus crewAttendanceStatus;
+    private final AttendanceStatus attendanceStatus;
+    private static final int LATE_STANDARD_MIN = 5;
+    private static final int ABSENCE_STANDARD_MIN = 30;
 
     public CrewAttendance(AttendanceTime attendanceTime) {
         this.attendanceTime = attendanceTime;
-        crewAttendanceStatus = new CrewAttendanceStatus(attendanceTime);
+        this.attendanceStatus = calculateAttendanceStatus(attendanceTime);
+    }
+
+    private AttendanceStatus calculateAttendanceStatus(AttendanceTime attendanceTime) {
+        int min = attendanceTime.minuteFromSchoolStartTime();
+        if (min > ABSENCE_STANDARD_MIN) {
+            return AttendanceStatus.ABSENCE;
+        }
+        if (min > LATE_STANDARD_MIN) {
+            return AttendanceStatus.LATE;
+        }
+        return AttendanceStatus.ATTENDANCE;
     }
 
     public AttendanceStatus attendanceStatus() {
-        return crewAttendanceStatus.attendanceStatus();
+        return attendanceStatus;
     }
 
     public LocalTime attendanceTime() {
@@ -35,12 +48,12 @@ public class CrewAttendance {
         }
         CrewAttendance that = (CrewAttendance) o;
         return Objects.equals(attendanceTime, that.attendanceTime) && Objects.equals(
-                crewAttendanceStatus, that.crewAttendanceStatus);
+                attendanceStatus, that.attendanceStatus);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(attendanceTime, crewAttendanceStatus);
+        return Objects.hash(attendanceTime, attendanceStatus);
     }
 }
